don't deploy remote state when using k3d provider

+++ src/mlstacks/cli/cli.py
@@ -26,6 +26,7 @@ from mlstacks.constants import (
     MLSTACKS_PACKAGE_NAME,
 )
 from mlstacks.enums import AnalyticsEventsEnum
+from mlstacks.models.stack import Stack
 from mlstacks.utils.cli_utils import (
     _get_spec_dir,
     confirmation,
@@ -45,7 +46,7 @@ from mlstacks.utils.terraform_utils import (
     get_stack_outputs,
     infracost_breakdown_stack,
 )
-from mlstacks.utils.yaml_utils import load_yaml_as_dict
+from mlstacks.utils.yaml_utils import load_stack_yaml, load_yaml_as_dict
 
 
 @click.group()
@@ -91,30 +92,30 @@ def deploy(
         debug (bool): Flag to enable debug mode to view raw Terraform logging
     """
     with analytics_client.EventHandler(AnalyticsEventsEnum.MLSTACKS_DEPLOY):
-        if not remote_state_bucket_name:
-            # generate random bucket name
-            letters = string.ascii_lowercase + string.digits
-            random_bucket_suffix = "".join(
-                random.choice(letters) for _ in range(6)  # noqa: S311
-            )
-            random_bucket_name = (
-                f"{DEFAULT_REMOTE_STATE_BUCKET_NAME}-{random_bucket_suffix}"
-            )
+        stack: Stack = load_stack_yaml(file)
+        if stack.provider.value != "k3d":
+            if remote_state_bucket_name:
+                deployed_bucket_url = remote_state_bucket_name
+                declare(f"Using '{deployed_bucket_url}' for remote state...")
+            else:
+                # generate random bucket name
+                letters = string.ascii_lowercase + string.digits
+                random_bucket_suffix = "".join(
+                    random.choice(letters) for _ in range(6)  # noqa: S311
+                )
+                random_bucket_name = f"{DEFAULT_REMOTE_STATE_BUCKET_NAME}-{random_bucket_suffix}"
 
-            # Remote state deployment
-            declare(
-                "Deploying remote state to bucket "
-                f"'{random_bucket_name}'...",
-            )
-            deployed_bucket_url = deploy_remote_state(
-                stack_path=file,
-                bucket_name=random_bucket_name,
-                debug_mode=debug,
-            )
-            declare("Remote state successfully deployed!")
-        else:
-            deployed_bucket_url = remote_state_bucket_name
-            declare(f"Using '{deployed_bucket_url}' for remote state...")
+                # Remote state deployment
+                declare(
+                    "Deploying remote state to bucket "
+                    f"'{random_bucket_name}'...",
+                )
+                deployed_bucket_url = deploy_remote_state(
+                    stack_path=file,
+                    bucket_name=random_bucket_name,
+                    debug_mode=debug,
+                )
+                declare("Remote state successfully deployed!")
 
         # Stack deployment
         declare(f"Deploying stack from '{file}'...")
@@ -203,17 +204,18 @@ def destroy(file: str, debug: bool = False, yes: bool = False) -> None:
             shutil.rmtree(tf_files_dir)
         declare(f"Stack '{stack_name}' has been destroyed.")
 
-        remote_state_dir = _get_remote_state_dir_path(provider)
-        if (
-            yes
-            or confirmation(
-                f"Would you like to destroy the Terraform remote state used "
-                f"for this stack on {provider}?",
-            )
-        ) and Path(remote_state_dir).exists():
-            destroy_remote_state(provider)
-            shutil.rmtree(remote_state_dir)
-        declare(f"Remote state for {provider} has been destroyed.")
+        if provider != "k3d":
+            remote_state_dir = _get_remote_state_dir_path(provider)
+            if (
+                yes
+                or confirmation(
+                    f"Would you like to destroy the Terraform remote state used "
+                    f"for this stack on {provider}?",
+                )
+            ) and Path(remote_state_dir).exists():
+                destroy_remote_state(provider)
+                shutil.rmtree(remote_state_dir)
+            declare(f"Remote state for {provider} has been destroyed.")
 
 
 @click.command()
